feat: refetch employees after upload
